move preprocessing functions to x86 pose class
